requirements.txt
langchain
langchain-community
pandas
boto3
python-dotenv
apache-airflow
requests
langchain-ollama
